+++ src/test/java/Homework17/BaseTest.java
@@ -1,0 +1,13 @@
+import io.github.bonigarcia.wdm.WebDriverManager;
+//import org.openqa.selenium.chrome.ChromeOptions;
+import org.testng.annotations.BeforeSuite;
+
+public class BaseTest {
+
+
+    @BeforeSuite
+    static void setupClass() {
+        WebDriverManager.chromedriver().setup();
+
+    }
+}
