+++ src/test/resources/TestNG.xml
@@ -1,13 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="TestNG">
+<suite name="Homework17">
 
-    <test name="Example Test" preserve-order="false">
+    <test name="Homework17" preserve-order="false">
         <classes>
             <class name="LoginTests"/>
             <class name="BaseTest"/>
-            <class name="Homework16"/>
             <class name="Homework17"/>
         </classes>
     </test>
